初期化処理のリファクタリング

+++ Assets/Scripts/Model/Goods/Cargo.cs
@@ -3,9 +3,9 @@
 namespace Model.Goods {
     public class Cargo {
         public GoodsEntity Goods { get; }
-        public int Amount { get; }
+        public float Amount { get; }
 
-        public Cargo(GoodsEntity goods, int amount) {
+        public Cargo(GoodsEntity goods, float amount) {
             Goods = goods;
             Amount = amount;
         }
