+++ Assets/Scripts/Model/Goods/ProduceAbility.cs
@@ -3,9 +3,9 @@
 namespace Model.Goods {
     public class ProduceAbility {
         public GoodsEntity OutputGoods { get; }
-        public int ProduceAmount { get; }
+        public float ProduceAmount { get; }
 
-        public ProduceAbility(GoodsEntity outputGoods, int produceAmount) {
+        public ProduceAbility(GoodsEntity outputGoods, float produceAmount) {
             OutputGoods = outputGoods;
             ProduceAmount = produceAmount;
         }
